+++ app/PersonioIntegration/Extensions.php
@@ -175,6 +175,7 @@ class Extensions {
 		// add extensions we deliver in this plugin.
 		$extension_list[] = '\PersonioIntegrationLight\PersonioIntegration\Availability';
 		$extension_list[] = '\PersonioIntegrationLight\PersonioIntegration\Manual_Import';
+		$extension_list[] = '\PersonioIntegrationLight\PersonioIntegration\Personio_Accounts';
 		$extension_list[] = '\PersonioIntegrationLight\PersonioIntegration\Show_Position_Xml';
 		$extension_list[] = '\PersonioIntegrationLight\PageBuilder\Page_Builders';