+++ app/PersonioIntegration/Position.php
@@ -829,24 +829,4 @@ class Position {
 			)
 		);
 	}
-
-	/**
-	 * Return the URL to edit this position on Personio.
-	 *
-	 * Needs the login URL to work.
-	 *
-	 * @return string
-	 */
-	public function get_edit_link_on_personio(): string {
-		// get the configured Personio Login URL.
-		$personio_login_url = get_option( 'personioIntegrationLoginUrl' );
-
-		// bail if no login URL is given.
-		if ( empty( $personio_login_url ) ) {
-			return '';
-		}
-
-		// get the URL.
-		return $personio_login_url . '/recruiting/position/' . $this->get_personio_id() . '?tab=job-details';
-	}
 }